Peliminary Execution
The execution folder was updated to include an error handler and database connector.
The system config was updated with test database username and password.
Placeholders were removed.
Changed order or Libraries so the connection library loads later, also changed the name to dbConnection.
Created a log library which is used by the new error handler. THey will be separated in to 'user' and 'system' files.
Created a log folder in sys/ to store logs.