Finalized Kore Workflow
The execution portion of workflow was finalized; now containing
all of its counter parts and error handling. Currently this
excludes text/stream, however a small update will allow for this
feature in the future.

The renderUI exec script was removed becuase it is not longer needed.
Now all requests will end up be handled by the api, erronoius requests
defaulting to the one required API module: UI. The UI module now
contains all UI handling.

Because UI handling is now done within the API, the processResult exec
script was moved from sys/500 to sys/900 so it is executed last. This
change will not require updates to session handling, but now requires
that sessions are handled prior to UI rendering.

The legacy rame.php was removed from ./ and replaced with kore.php to
bring current with this project name and clone. The htaccess file was
updated to reflect this change.

The UI API was updated to now build headers. THe system config was 
updated with assets to render in to the html head section of the
output document when requests are in text/html.