New Assets
UI ./assets were updated to include bootstrap (local copy), and jQuery (linked).
The default theme for bootstrap will be used for now. The system settings 
(./sys/settings/system.conf) was updated to include the new assets. 

For compatability reasons, the script folder was changed from 'script' to
'js', and this is also reflected in the system settings.