API Additions
A new API module was added (authoority) to handle permissions
and authentication. It will interact with the session to determine
available functions.

Some test functions have been added for now to build a working model.
The authority module is currently is its begining stages.

The API library (./sys/lib/200/api.lib) was updated to allow post data
to be passed directly through to a function when type is POST.

An additional update to the API library allows for direct function calls
on objects. The previous concern with this was permissions and trying 
to build a complex platform to handle what functions can be called on
what objects. In order to make the API truley dynamic, the development
should not be slowed down by adding complex permissions parameters; so
a very dyanmic system for permissions was added to the api interface
and public functions of objects in the api can now be called directly
from the api restfully and with a robust permissions layer. As a side-note;
because the API interface was updated, all affected class objects were
subsequently updated with the appropriate changes.

The ./ui folder was updated to include a landing page for 'authenticated'
members. This is a temporary landing ground for testing.