Execution Updates
Session handling was modified to work better with the Authority API. It
can now be modified instead of being destroyed and recreated, avoiding
an unneeded page reload.

The processRequest was updated to include the new permissions layer for
the API. Currently, the instansiated object is used to provide it's own
permissions, and if false is returned, a fatal user error (2048) is thrown.