UI and API Updates
The UI and API have both been updated with changes
that help API components provide UI elements that
may be needed when creating specific plugins.

The first part of this is the server-side API structure
which has been updated to include ./pages/ and settgins.cong
in the API plugin folder to allow some low-level processing
to happen when plugins are used. The most complete portion being
able to load UI elements from the API.

The JavaScript libraries were updated and the original rame()
object was modified to accomplish this. API UI elements stored
as .phtml files now be called and loaded to the UI (async).